feat(*) refactor data service
